Обновли иерархию папок
