--- Account-microservice/src/main/java/com/example/Account_microservice/CustomerUserDetailsService.java
@@ -1,21 +1,0 @@
-package com.example.Account_microservice;
-
-
-import com.example.Account_microservice.user.serivice.UserService;
-import lombok.RequiredArgsConstructor;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UserDetailsService;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-@RequiredArgsConstructor
-public class CustomerUserDetailsService implements UserDetailsService {
-
-    private final UserService userService;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return  userService.findUser(username);
-    }
-}
