+++ Account-microservice/src/main/java/com/example/Account_microservice/exception/BadRequestExceptionCustomer.java
@@ -1,0 +1,9 @@
+package com.example.Account_microservice.exception;
+
+
+import java.util.List;
+
+
+public record BadRequestExceptionCustomer(
+        List<Validate> errors)  {
+}
