+++ Account-microservice/src/main/java/com/example/Account_microservice/exception/Validate.java
@@ -1,0 +1,4 @@
+package com.example.Account_microservice.exception;
+
+public record Validate(String error) {
+}
