--- Account-microservice/src/main/java/com/example/Account_microservice/exeption/BadRequestExceptionCustomer.java
@@ -1,9 +1,0 @@
-package com.example.Account_microservice.exeption;
-
-
-import java.util.List;
-
-
-public record BadRequestExceptionCustomer(
-        List<Validate> errors)  {
-}
