--- Account-microservice/src/main/java/com/example/Account_microservice/exeption/Validate.java
@@ -1,4 +1,0 @@
-package com.example.Account_microservice.exeption;
-
-public record Validate(String error) {
-}
