--- Account-microservice/src/main/java/com/example/Account_microservice/jwt/component/JwtTokenIntrospector.java
@@ -1,7 +1,0 @@
-package com.example.Account_microservice.jwt.component;
-
-import com.example.Account_microservice.jwt.dto.JwtTokenIntrospectResponse;
-
-public interface JwtTokenIntrospector {
-    JwtTokenIntrospectResponse build(String token);
-}
