--- Account-microservice/src/main/java/com/example/Account_microservice/jwt/dto/JwtAuthenticationResponse.java
@@ -1,6 +1,0 @@
-package com.example.Account_microservice.jwt.dto;
-
-public record JwtAuthenticationResponse(
-        String token
-) {
-}
