--- Account-microservice/src/main/java/com/example/Account_microservice/jwt/dto/JwtAuthority.java
@@ -1,6 +1,0 @@
-package com.example.Account_microservice.jwt.dto;
-
-public record JwtAuthority(
-        String authority
-) {
-}
