--- Account-microservice/src/main/java/com/example/Account_microservice/jwt/service/JwtExtractService.java
@@ -1,18 +1,0 @@
-package com.example.Account_microservice.jwt.service;
-
-import com.example.Account_microservice.jwt.dto.JwtAuthority;
-import com.example.Account_microservice.user.model.Role;
-
-import java.util.Date;
-import java.util.List;
-import java.util.Set;
-
-public interface JwtExtractService {
-    Long extractUserId(String token);
-    String extractUserName(String token);
-    String extractSubject(String token);
-    Date extractIssuedAt(String token);
-    Date extractExpiration(String token);
-    Boolean isTokenExpired(String token);
-    List<JwtAuthority> extractRole(String token);
-}
