--- Account-microservice/src/main/java/com/example/Account_microservice/jwt/service/JwtService.java
@@ -1,15 +1,0 @@
-package com.example.Account_microservice.jwt.service;
-
-import org.springframework.security.core.userdetails.UserDetails;
-
-public interface JwtService {
-
-    String generateToken(UserDetails userDetails);
-
-
-    boolean isTokenValid(String token, UserDetails userDetails);
-
-    Long getExpirationTime(String token);
-
-
-}
