+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/black_list/repository/BlackListRepository.java
@@ -1,0 +1,11 @@
+package com.example.Account_microservice.security.jwt.black_list.repository;
+
+import com.example.Account_microservice.security.jwt.black_list.model.BlackListToken;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+
+
+public interface BlackListRepository extends JpaRepository<BlackListToken, Long> {
+
+    Boolean existsByToken(String token);
+}
