+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/black_list/service/BlackListTokenService.java
@@ -1,0 +1,14 @@
+package com.example.Account_microservice.security.jwt.black_list.service;
+
+import com.example.Account_microservice.security.jwt.black_list.model.BlackListToken;
+
+import java.util.List;
+
+public interface BlackListTokenService {
+
+    void save(BlackListToken token);
+
+    Boolean isTokenBlacklisted(String token);
+
+    List<BlackListToken> findAll();
+}
