+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/component/JwtTokenIntrospector.java
@@ -1,0 +1,7 @@
+package com.example.Account_microservice.security.jwt.component;
+
+import com.example.Account_microservice.security.jwt.dto.JwtTokenIntrospectResponse;
+
+public interface JwtTokenIntrospector {
+    JwtTokenIntrospectResponse build(String token);
+}
