+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/dto/JwtAuthenticationResponse.java
@@ -1,0 +1,6 @@
+package com.example.Account_microservice.security.jwt.dto;
+
+public record JwtAuthenticationResponse(
+        String token
+) {
+}
