+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/dto/JwtAuthority.java
@@ -1,0 +1,6 @@
+package com.example.Account_microservice.security.jwt.dto;
+
+public record JwtAuthority(
+        String authority
+) {
+}
