+++ Account-microservice/src/main/java/com/example/Account_microservice/security/jwt/service/JwtService.java
@@ -1,0 +1,15 @@
+package com.example.Account_microservice.security.jwt.service;
+
+import org.springframework.security.core.userdetails.UserDetails;
+
+public interface JwtService {
+
+    String generateToken(UserDetails userDetails);
+
+
+    boolean isTokenValid(String token, UserDetails userDetails);
+
+    Long getExpirationTime(String token);
+
+
+}
